Update counter_ebx.asm
Explanation of EBX Counter Loop

The EBX register is utilized as a counter in this program, with a value of 0.  EBX is continuously increased by 1 inside the loop.  The software uses the cmp instruction to compare EBX to 10 after incrementing.  The software returns to the beginning of the loop and counts again if EBX is less than 10.

 The loop finishes when EBX hits 10, at which point this process is repeated.  We can see that EBX rises from 0 to 9 during the loop iterations when we use GDB to debug the program.  As soon as EBX equals 10, the loop ends and the program ends.

Counters can be built in assembly without depending on the loop instruction thanks to this manual looping technique that uses EBX and conditional jump (jl).  It demonstrates in detail how comparison instructions and registers cooperate to regulate program flow.